+++ test/node/protocol/blockchain_test.go
@@ -9,34 +9,6 @@ import (
 	"testing"
 )
 
-//func Test_IsEmpty_Empty_ReturnsTrue(t *testing.T) {
-//	blockchain := protocol.NewBlockchain(0, nil, nil)
-//	isEmpty := blockchain.IsEmpty()
-//	test.Assert(t, isEmpty, "blockchain should be empty")
-//}
-//
-//func Test_IsEmpty_BlockAdded_ReturnsFalse(t *testing.T) {
-//	blockchain := protocol.NewBlockchain(0, nil, nil)
-//	blockchain.AddBlock(protocol.NewBlock(0, [32]byte{}, nil, nil))
-//	isEmpty := blockchain.IsEmpty()
-//	test.Assert(t, !isEmpty, "blockchain should not be empty")
-//}
-//
-//func Test_BlockResponses_Empty_ReturnsNil(t *testing.T) {
-//	blockchain := protocol.NewBlockchain(0, nil, nil)
-//	blockResponses := blockchain.Blocks()
-//	test.Assert(t, blockResponses == nil, "block responses should be nil")
-//}
-//
-//func Test_CalculateTotalAmount__ReturnsTotalAmount(t *testing.T) {
-//	blockchain := protocol.NewBlockchain(0, nil, nil)
-//	transaction := protocol.NewTransaction("recipient", "sender", nil, 0, 10)
-//	block := protocol.NewBlock(0, [32]byte{}, []*protocol.Transaction{transaction}, nil)
-//	blockchain.AddBlock(block)
-//	actualTotalAmount := blockchain.CalculateTotalAmount(0, "recipient")
-//	var expectedTotalAmount uint64 = 10
-//	test.Assert(t, actualTotalAmount == expectedTotalAmount, fmt.Sprintf("Wrong total amount. Expected: %d - Actual: %d", expectedTotalAmount, actualTotalAmount))
-//}
 func Test_Verify_NeighborBlockchainIsBetter_IsReplaced(t *testing.T) {
 	registrable := NewRegistrableMock()
 	IsRegisteredMock = func(address string) (bool, error) { return true, nil }
