+++ test/node/protocol/pool_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/my-cloud/ruthenium/src/node/protocol"
 	"github.com/my-cloud/ruthenium/src/ui/server"
 	"github.com/my-cloud/ruthenium/test"
-	"github.com/my-cloud/ruthenium/test/clock"
+	"github.com/my-cloud/ruthenium/test/node"
 	"testing"
 	"time"
 )
@@ -23,7 +23,7 @@ func Test_AddTransaction_Allowed_TransactionAdded(t *testing.T) {
 	walletBAddress := walletB.Address()
 	registrable := NewRegistrableMock()
 	IsRegisteredMock = func(address string) (bool, error) { return true, nil }
-	watch := clock.NewWatch()
+	watch := node.NewWatch()
 	validationTimer := time.Nanosecond
 	logger := log.NewLogger(log.Fatal)
 	blockchain := protocol.NewBlockchain(registrable, validationTimer, watch, logger)
@@ -62,7 +62,7 @@ func Test_AddTransaction_NotAllowed_TransactionNotAdded(t *testing.T) {
 	walletBAddress := walletB.Address()
 	registrable := NewRegistrableMock()
 	IsRegisteredMock = func(address string) (bool, error) { return true, nil }
-	watch := clock.NewWatch()
+	watch := node.NewWatch()
 	validationTimer := time.Nanosecond
 	logger := log.NewLogger(log.Fatal)
 	blockchain := protocol.NewBlockchain(registrable, validationTimer, watch, logger)
